Add Property to Demo code blocks. Fix various layout issues.
Getting things to align in the demo is getting unwieldy with Code’s old
fixed 24px margin. You can now set margins as you wish. This change
touches most of the demo, because we have to re-install the 24px margin
in a number of places.

This commit also makes a number of minor adjustments to demo layout.