Add support for custom event handlers on Textfield
This also refactors the way event handlers are used internally.
Instead of adding the internal handlers on the input element.
The handlers are now added to the containing div using events that
are known to bubble (input, focusin, focusout)

The user event handlers are now added to the input element.