Make Menu apply styling to top-level div; position same
In the interest of controlling placement of the menu, styling supplied
to menu is now applied to the top-level div.

The top-level div is also styled `position: relative;` to fix placement
issues, and the activating menu button now closes the menu when pressed
if its active to work around yet more bugs due to bad
getBoundingClientRect.

Fix #143.