Refactor Chip API
New Chip API uses opaque type to ensure that only specific
types are allowed when creating a chip.

Add ability to add event handlers to Chip and also the
delete action has been added as a property on the Chip itself
rather than provided as Content. This is because there should really
only be a single delete action and it does not really need other functionality
other than responding to either as a link or to clicks.