Initial implementation of generic attribute support
This is a initial implementation of generic attribute support
to enable arbitrary `Html.Attributes`. This implementation
tries to minimize the likelyhood of overriding internally used
attributes while still maintaining `MINOR` change.

This implementation may be revisited in the next MAJOR version
to update the implementation.

Add namespaced qualifiers to `Internal.attribute`,
this way internal attributes are easy to search for as well separate
them from `Options.attribute`