Make Temporary and Persistent Drawer programatic
It is implemented the same way as Dialog:
* Drawer.open must passed to properties of Drawer.view
* for temporary drawer, an `onClose` property must be provided to handle
  the closing of the drawer

There is no more subscriptions in Drawer's implementation in order to
reduce code complexity in state handling (the subscriptions does not
have access to the `onClose` message).Thus the `onClose` message is now
only sent when the user clicks on the backdrop of the drawer.

It is now possible to close the drawer programatically by adding or
removing the `Drawer.open` property on `the Drawer.view` options.
This can be useful to close the drawer when the user clicks on a
drawer's item.